requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
pyephem==3.7.6.0
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
